Split character-specific item tables
