+++ src/melee/it/it_279C.c
@@ -2,7 +2,6 @@
 
 #include "it_2725.h"
 #include "ithitbox.h"
-#include "itzako.h"
 
 #include "baselib/random.h"
 #include "db/db.h"
@@ -11,128 +10,40 @@
 #include "ft/ft_0892.h"
 #include "ft/ftlib.h"
 #include "gm/gm_unsplit.h"
+#include "it/inlines.h"
 #include "it/itdraw.h"
-#include "it/items/it_2ADA.h"
-#include "it/items/it_2E5A.h"
-#include "it/items/it_2F28.h"
-#include "it/items/itarwinglaser.h"
 #include "it/items/itcerebi.h"
 #include "it/items/itchicorita.h"
-#include "it/items/itclimbersblizzard.h"
-#include "it/items/itclimbersice.h"
-#include "it/items/itclimbersstring.h"
-#include "it/items/itclinkmilk.h"
-#include "it/items/itcrazyhandbomb.h"
-#include "it/items/itdrmariopill.h"
 #include "it/items/itentei.h"
 #include "it/items/itfire.h"
-#include "it/items/itfoxblaster.h"
-#include "it/items/itfoxlaser.h"
 #include "it/items/itfreezer.h"
 #include "it/items/itfushigibana.h"
-#include "it/items/itgamewatchbreath.h"
-#include "it/items/itgamewatchchef.h"
-#include "it/items/itgamewatchfire.h"
-#include "it/items/itgamewatchgreenhouse.h"
-#include "it/items/itgamewatchjudge.h"
-#include "it/items/itgamewatchmanhole.h"
-#include "it/items/itgamewatchpanic.h"
-#include "it/items/itgamewatchparachute.h"
-#include "it/items/itgamewatchrescue.h"
-#include "it/items/itgamewatchturtle.h"
-#include "it/items/itgreatfoxlaser.h"
 #include "it/items/ithassam.h"
-#include "it/items/itheiho.h"
 #include "it/items/ithinoarashi.h"
 #include "it/items/ithitodeman.h"
 #include "it/items/ithouou.h"
 #include "it/items/itkabigon.h"
 #include "it/items/itkamex.h"
-#include "it/items/itkirbycutterbeam.h"
-#include "it/items/itkirbygamewatchchefpan.h"
-#include "it/items/itkirbyhammer.h"
 #include "it/items/itkireihana.h"
-#include "it/items/itklap.h"
-#include "it/items/itkoopaflame.h"
-#include "it/items/itkyasarin.h"
-#include "it/items/itkyasarinegg.h"
-#include "it/items/itleadead.h"
-#include "it/items/itlikelike.h"
-#include "it/items/itlinkbomb.h"
-#include "it/items/itlinkbow.h"
-#include "it/items/itlinkhookshot.h"
 #include "it/items/itlizardon.h"
 #include "it/items/itlucky.h"
 #include "it/items/itlugia.h"
-#include "it/items/itluigifireball.h"
 #include "it/items/itmaril.h"
-#include "it/items/itmariocape.h"
-#include "it/items/itmariofireball.h"
 #include "it/items/itmarumine.h"
-#include "it/items/itmasterhandbullet.h"
-#include "it/items/itmasterhandlaser.h"
 #include "it/items/itmatadogas.h"
-#include "it/items/itmato.h"
 #include "it/items/itmetamon.h"
 #include "it/items/itmew.h"
-#include "it/items/itmewtwodisable.h"
-#include "it/items/itmewtwoshadowball.h"
-#include "it/items/itnessbat.h"
-#include "it/items/itnesspkfire.h"
-#include "it/items/itnesspkfirepillar.h"
-#include "it/items/itnesspkflash.h"
-#include "it/items/itnesspkflashexplode.h"
-#include "it/items/itnesspkthunderball.h"
-#include "it/items/itnesspkthundertrail.h"
-#include "it/items/itnessyoyo.h"
-#include "it/items/itnokonoko.h"
-#include "it/items/itoctarock.h"
-#include "it/items/itoctarockstone.h"
 #include "it/items/itoldkuri.h"
-#include "it/items/itoldottosea.h"
-#include "it/items/itpatapata.h"
-#include "it/items/itpeachexplode.h"
-#include "it/items/itpeachparasol.h"
-#include "it/items/itpeachtoad.h"
-#include "it/items/itpeachtoadspore.h"
-#include "it/items/itpeachturnip.h"
-#include "it/items/itpikachuthunder.h"
-#include "it/items/itpikachutjoltair.h"
-#include "it/items/itpikachutjoltground.h"
 #include "it/items/itpippi.h"
 #include "it/items/itporygon2.h"
 #include "it/items/itraikou.h"
-#include "it/items/itsamusbomb.h"
-#include "it/items/itsamuschargeshot.h"
-#include "it/items/itsamusgrapple.h"
-#include "it/items/itsamusmissile.h"
-#include "it/items/itseakchain.h"
-#include "it/items/itseakneedleheld.h"
-#include "it/items/itseakneedlethrown.h"
-#include "it/items/itseakvanish.h"
 #include "it/items/itsonans.h"
 #include "it/items/itsuikun.h"
 #include "it/items/itthunder.h"
-#include "it/items/ittincle.h"
 #include "it/items/ittogepy.h"
-#include "it/items/ittools.h"
 #include "it/items/ittosakinto.h"
 #include "it/items/itunknown.h"
-#include "it/items/itwhispyapple.h"
-#include "it/items/itwhitebea.h"
-#include "it/items/ityaku.h"
-#include "it/items/ityoshieggthrow.h"
-#include "it/items/ityoshistar.h"
-#include "it/items/itzeldadinfire.h"
-#include "it/items/itzeldadinfireexplode.h"
-#include "it/items/itzgshell.h"
-#include "it/items/itzrshell.h"
 #include "it/itgroundcoll.h"
-#include "items/itcoin.h"
-#include "items/itfoxillusion.h"
-#include "items/itlinkarrow.h"
-#include "items/itlinkboomerang.h"
-#include "items/ityoshiegglay.h"
 #include "lb/lbvector.h"
 #include "pl/plattack.h"
 #include "pl/plbonuslib.h"
@@ -1015,2690 +926,6 @@ static ItemKind common_pokemon[] = {
     It_PKind_Pippi,     It_PKind_Togepy,      It_PKind_Hitodeman,
     It_PKind_Lucky,     It_PKind_Porygon2,    It_PKind_Hinoarashi,
     It_PKind_Maril,     It_PKind_Fushigibana, It_PKind_Terminate,
-};
-
-/// Character-specific items
-struct sdata_ItemGXLink it_803F2F28[118] ATTRIBUTE_ALIGN(8) = {
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8029CD18, it_8029CD18,
-
-    it_8026EECC, it_8026EECC, it_8029FDDC, it_8029FDDC, it_8026EECC,
-    it_8026EECC, it_802A7D8C, it_802A7D8C, it_8026EECC, it_8026EECC,
-
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-
-    it_8026EECC, it_8026EECC, it_802A7D8C, it_802A7D8C, it_8026EECC,
-
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-
-    it_8026EECC, it_802F13B4, it_8026EECC,
-};
-
-struct ItemLogicTable it_803F3100[ARRAY_SIZE(it_803F2F28)] = {
-    {
-        // Kuriboh
-        it_803F8320,
-        it_802D73F0,
-        it_2725_Logic9_Destroyed,
-        itOldKuri_Logic0_PickedUp,
-        it_2725_Logic0_Dropped,
-        it_2725_Logic0_Thrown,
-        NULL,
-        it_2725_Logic0_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Leadead
-        it_803F8EB0,
-        it_802E8BCC,
-        itLeadead_Logic1_Destroyed,
-        itLeadead_Logic1_PickedUp,
-        itLeadead_Logic1_Dropped,
-        itLeadead_Logic1_Thrown,
-        NULL,
-        it_802E8CD8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Octarock
-        it_803F8B48,
-        it_802E4A44,
-        it_2725_Logic9_Destroyed,
-        itOctarock_Logic2_PickedUp,
-        itOctarock_Logic2_Dropped,
-        itOctarock_Logic2_Thrown,
-        NULL,
-        it_802E4B00,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802E5AA4,
-    },
-    {
-        // Ottosea
-        it_803F89C8,
-        it_802E2470,
-        it_2725_Logic3_Destroyed,
-        it_802E31F8,
-        it_802E32B4,
-        it_802E3314,
-        NULL,
-        it_2725_Logic8_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802E3784,
-    },
-    {
-        // Octarock stone
-        it_803F8E90,
-        itOctarockstone_802E878C,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itOctarockstone_Logic4_DmgDealt,
-        NULL,
-        NULL,
-        itOctarockstone_Logic4_Reflected,
-        itOctarockstone_Logic4_Clanked,
-        itOctarockstone_Logic4_Absorbed,
-        itOctarockstone_Logic4_ShieldBounced,
-        itOctarockstone_Logic4_HitShield,
-        itOctarockstone_802E89B0,
-    },
-    {
-        // Mario fireball
-        it_803F6788,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itMarioFireball_Logic87_DmgDealt,
-        NULL,
-        NULL,
-        itMarioFireball_Logic87_Reflected,
-        itMarioFireball_Logic87_Clanked,
-        itMarioFireball_Logic87_Absorbed,
-        itMarioFireball_Logic87_ShieldBounced,
-        itMarioFireball_Logic87_HitShield,
-        itMarioFireball_Logic87_EvtUnk,
-    },
-    {
-        // Dr Mario pill
-        it_803F75D0,
-        NULL,
-        NULL,
-        itDrMarioPill_PickedUp,
-        NULL,
-        NULL,
-        itDrMarioPill_DmgDealt,
-        NULL,
-        NULL,
-        itDrMarioPill_Reflected,
-        itDrMarioPill_Clanked,
-        itDrMarioPill_Absorbed,
-        itDrMarioPill_ShieldBounced,
-        itDrMarioPill_HitShield,
-        itDrMarioPill_EvtUnk,
-    },
-    {
-        // Kirby cutter beam
-        it_803F6798,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itKirbyCutterBeam_Logic7_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic7_Reflected,
-        itKirbyCutterBeam_Logic7_Clanked,
-        itKirbyCutterBeam_Logic7_Absorbed,
-        it_2725_Logic7_ShieldBounced,
-        itKirbyCutterBeam_Logic7_HitShield,
-        it_8029C4B4,
-    },
-    {
-        // Kirby hammer
-        it_803F6C98,
-        NULL,
-        itKirbyHammer_Logic8_Destroyed,
-        itKirbyHammer_Logic8_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802ADDB0,
-    },
-    {
-        it_803F6C88,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802ADBE4,
-    },
-    {
-        it_803F9470,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802F2BDC,
-    },
-    {
-        // Fox laser
-        it_803F67D0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itFoxLaser_Logic94_Clanked,
-        NULL,
-        NULL,
-        itFoxLaser_Logic94_Reflected,
-        itFoxLaser_Logic94_Clanked,
-        itFoxLaser_Logic94_Absorbed,
-        itFoxLaser_Logic94_ShieldBounced,
-        itFoxLaser_Logic94_HitShield,
-        itFoxLaser_Logic94_EvtUnk,
-    },
-    {
-        // Falco laser
-        it_803F67D0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itFoxLaser_Logic94_Clanked,
-        NULL,
-        NULL,
-        itFoxLaser_Logic94_Reflected,
-        itFoxLaser_Logic94_Clanked,
-        itFoxLaser_Logic94_Absorbed,
-        itFoxLaser_Logic94_ShieldBounced,
-        itFoxLaser_Logic94_HitShield,
-        itFoxLaser_Logic94_EvtUnk,
-    },
-    {
-        // Fox illusion
-        it_803F6818,
-        NULL,
-        itFoxIllusion_Logic14_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        itFoxIllusion_Logic14_DmgDealt,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_8029D948,
-    },
-    {
-        // Falco phantasm
-        it_803F6818,
-        NULL,
-        itFoxIllusion_Logic14_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        itFoxIllusion_Logic14_DmgDealt,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_8029D948,
-    },
-    {
-        // Link bomb
-        it_803F6888,
-        NULL,
-        NULL,
-        it_8029DEB0,
-        it_8029EC34,
-        it_8029E5D0,
-        it_8029F960,
-        itLinkBomb_Logic16_DmgReceived,
-        itLinkBomb_Logic16_EnteredAir,
-        itLinkBomb_Logic16_Reflected,
-        NULL,
-        NULL,
-        itLinkBomb_Logic16_ShieldBounced,
-        itLinkBomb_Logic16_HitShield,
-        it_8029FD84,
-    },
-    {
-        // Young Link bomb
-        it_803F6888,
-        NULL,
-        NULL,
-        it_8029DEB0,
-        it_8029EC34,
-        it_8029E5D0,
-        it_8029F960,
-        itLinkBomb_Logic16_DmgReceived,
-        itLinkBomb_Logic16_EnteredAir,
-        itLinkBomb_Logic16_Reflected,
-        NULL,
-        NULL,
-        itLinkBomb_Logic16_ShieldBounced,
-        itLinkBomb_Logic16_HitShield,
-        it_8029FD84,
-    },
-    {
-        // Link boomerang
-        it_803F6920,
-        NULL,
-        itLinkBoomerang_Logic18_Destroyed,
-        it_802A0E70,
-        it_802A0F84,
-        it_802A0F88,
-        it_802A1F08,
-        NULL,
-        NULL,
-        it_802A20E8,
-        it_802A1FA8,
-        itLinkBoomerang_Logic18_Absorbed,
-        it_802A2320,
-        it_802A2288,
-        it_802A23CC,
-    },
-    {
-        // Young Link boomerang
-        it_803F6920,
-        NULL,
-        itLinkBoomerang_Logic18_Destroyed,
-        it_802A0E70,
-        it_802A0F84,
-        it_802A0F88,
-        it_802A1F08,
-        NULL,
-        NULL,
-        it_802A20E8,
-        it_802A1FA8,
-        itLinkBoomerang_Logic18_Absorbed,
-        it_802A2320,
-        it_802A2288,
-        it_802A23CC,
-    },
-    {
-        // Link hookshot
-        it_803F6998,
-        it_802A2418,
-        NULL,
-        itLinkHookshot_Logic20_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802A7D40,
-    },
-    {
-        // Young Link hookshot
-        it_803F6998,
-        NULL,
-        NULL,
-        itLinkHookshot_Logic20_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802A7D40,
-    },
-    {
-        // Link arrow
-        it_803F6A28,
-        NULL,
-        itLinkArrow_Logic98_Destroyed,
-        itLinkArrow_Logic98_PickedUp,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_DmgDealt,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_Reflected,
-        itLinkArrow_Logic98_Clanked,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_HitShield,
-        itLinkArrow_Logic98_EvtUnk,
-    },
-    {
-        // Young Link arrow
-        it_803F6A28,
-        NULL,
-        itLinkArrow_Logic98_Destroyed,
-        itLinkArrow_Logic98_PickedUp,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_DmgDealt,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_Reflected,
-        itLinkArrow_Logic98_Clanked,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_HitShield,
-        itLinkArrow_Logic98_EvtUnk,
-    },
-    {
-        // Ness PK fire
-        it_803F6B28,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic23_DmgDealt,
-        NULL,
-        NULL,
-        itNessPKFire_Logic23_Reflected,
-        it_2725_Logic23_Clanked,
-        itNessPKFire_Logic23_Absorbed,
-        it_2725_Logic23_ShieldBounced,
-        itNessPKFire_Logic23_HitShield,
-        it_802AA474,
-    },
-    {
-        // Ness PK fire pillar
-        it_803F6B60,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itNesspkfirepillar_Logic24_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itNesspkfirepillar_Logic24_EvtUnk,
-    },
-    {
-        // Ness PK flash
-        it_803F6B70,
-        NULL,
-        it_2725_Logic102_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itNesspkflash_Logic102_Reflected,
-        itNesspkflash_Logic102_Clanked,
-        itNesspkflash_Logic102_Absorbed,
-        NULL,
-        NULL,
-        itNesspkflash_Logic102_EvtUnk,
-    },
-    {
-        // Ness PK thunder ball
-        it_803F6BC8,
-        NULL,
-        it_802AB90C,
-        NULL,
-        NULL,
-        NULL,
-        itNessPKThunderball_Logic26_DmgDealt,
-        NULL,
-        NULL,
-        it_802AC098,
-        it_802AC074,
-        it_802AC338,
-        it_802AC35C,
-        it_802AC3F8,
-        it_802AC41C,
-    },
-    {
-        // Ness PK thunder trail 1
-        it_803F6C08,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Ness PK thunder trail 2
-        it_803F6C08,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Ness PK thunder trail 3
-        it_803F6C08,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Ness PK thunder trail 4
-        it_803F6C08,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Fox blaster
-        it_803F6CA8,
-        NULL,
-        NULL,
-        itFoxBlaster_Logic96_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itFoxBlaster_Logic96_EvtUnk,
-    },
-    {
-        // Falco blaster
-        it_803F6CA8,
-        NULL,
-        NULL,
-        itFoxBlaster_Logic96_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itFoxBlaster_Logic96_EvtUnk,
-    },
-    {
-        // Link bow
-        it_803F6E98,
-        NULL,
-        itLinkBow_Logic100_Destroyed,
-        itLinkBow_Logic100_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itLinkBow_Logic100_EvtUnk,
-    },
-    {
-        // Young Link bow
-        it_803F6E98,
-        NULL,
-        itLinkBow_Logic100_Destroyed,
-        itLinkBow_Logic100_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itLinkBow_Logic100_EvtUnk,
-    },
-    {
-        // Ness PK flush explode
-        it_803F6F40,
-        NULL,
-        it_2725_Logic103_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itNessPKFlashExplode_Logic103_Clanked,
-        itNessPKFlashExplode_Logic103_Absorbed,
-        itNessPKFlashExplode_Logic103_ShieldBounced,
-        itNessPKFlashExplode_Logic103_HitShield,
-        itNessPKFlashExplode_Logic103_EvtUnk,
-    },
-    {
-        // Sheik needle thrown
-        it_803F6F50,
-        NULL,
-        itSeakNeedleThrown_Logic109_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic109_DmgDealt,
-        it_2725_Logic109_DmgReceived,
-        NULL,
-        it_2725_Logic109_Reflected,
-        it_2725_Logic109_Clanked,
-        NULL,
-        it_2725_Logic109_ShieldBounced,
-        it_2725_Logic109_HitShield,
-        itSeakNeedleThrown_Logic109_EvtUnk,
-    },
-    {
-        // Sheik needle held
-        it_803F70A8,
-        NULL,
-        NULL,
-        itSeakNeedleHeld_Logic110_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itSeakNeedleHeld_Logic110_EvtUnk,
-    },
-    {
-        // Pikachu thunder
-        it_803F70C8,
-        NULL,
-        it_2725_Logic39_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        itPikachuThunder_Logic39_DmgDealt,
-        NULL,
-        NULL,
-        NULL,
-        itPikachuThunder_Logic39_Clanked,
-        itPikachuThunder_Logic39_Absorbed,
-        NULL,
-        itPikachuThunder_Logic39_HitShield,
-        itPikachuThunder_Logic39_EvtUnk,
-    },
-    {
-        // Pichu thunder
-        it_803F70C8,
-        NULL,
-        it_2725_Logic39_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        itPikachuThunder_Logic39_DmgDealt,
-        NULL,
-        NULL,
-        NULL,
-        itPikachuThunder_Logic39_Clanked,
-        itPikachuThunder_Logic39_Absorbed,
-        NULL,
-        itPikachuThunder_Logic39_HitShield,
-        itPikachuThunder_Logic39_EvtUnk,
-    },
-    {
-        // Mario cape
-        it_803F70F8,
-        NULL,
-        itMarioCape_Logic41_Destroyed,
-        it_2725_Logic41_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802B2870,
-    },
-    {
-        // Dr Mario cape
-        it_803F70F8,
-        NULL,
-        itMarioCape_Logic41_Destroyed,
-        it_2725_Logic41_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802B2870,
-    },
-    {
-        // Sheik vanish
-        it_803F70B8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itSeakVanish_Logic42_DmgDealt,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802B1DCC,
-    },
-    {
-        // Yoshi egg throw
-        it_803F7118,
-        NULL,
-        NULL,
-        itYoshiEggThrow_Logic43_PickedUp,
-        NULL,
-        NULL,
-        it_802B2C04,
-        NULL,
-        NULL,
-        it_802B2E5C,
-        it_2725_Logic43_Clanked,
-        NULL,
-        it_802B2F88,
-        it_802B2E7C,
-        it_802B2FA8,
-    },
-    {
-        // Yoshi egg lay
-        it_803F94A0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_27CF_Logic114_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_27CF_Logic114_EvtUnk,
-    },
-    {
-        // Yoshi star
-        it_803F7158,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802B309C,
-        NULL,
-        NULL,
-        it_802B314C,
-        it_802B30C0,
-        it_802B3108,
-        it_802B312C,
-        it_802B30E4,
-        it_802B3348,
-    },
-    {
-        // Pikachu tjolt ground
-        it_803F7190,
-        NULL,
-        it_2725_Logic106_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic106_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic106_Reflected,
-        it_2725_Logic106_Clanked,
-        it_2725_Logic106_Absorbed,
-        it_2725_Logic106_ShieldBounced,
-        it_2725_Logic106_HitShield,
-        itPikachuTJoltGround_Logic106_EvtUnk,
-    },
-    {
-        // Pikachu tjolt air
-        it_803F71D8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic107_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic107_Reflected,
-        it_2725_Logic107_Clanked,
-        it_2725_Logic107_Absorbed,
-        it_2725_Logic107_ShieldBounced,
-        it_2725_Logic107_HitShield,
-        itPikachuTJoltAir_Logic107_EvtUnk,
-    },
-    {
-        // Pichu tjolt ground
-        it_803F7190,
-        NULL,
-        it_2725_Logic106_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic106_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic106_Reflected,
-        it_2725_Logic106_Clanked,
-        it_2725_Logic106_Absorbed,
-        it_2725_Logic106_ShieldBounced,
-        it_2725_Logic106_HitShield,
-        itPikachuTJoltGround_Logic106_EvtUnk,
-    },
-    {
-        // Pichu tjolt air
-        it_803F71D8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic107_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic107_Reflected,
-        it_2725_Logic107_Clanked,
-        it_2725_Logic107_Absorbed,
-        it_2725_Logic107_ShieldBounced,
-        it_2725_Logic107_HitShield,
-        itPikachuTJoltAir_Logic107_EvtUnk,
-    },
-    {
-        // Samus bomb
-        it_803F7220,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itSamusBomb_Logic50_DmgDealt,
-        NULL,
-        itSamusBomb_Logic50_EnteredAir,
-        it_2725_Logic50_Reflected,
-        itSamusBomb_Logic50_Clanked,
-        NULL,
-        itSamusBomb_Logic50_ShieldBounced,
-        itSamusBomb_Logic50_HitShield,
-        itSamusBomb_Logic50_EvtUnk,
-    },
-    {
-        // Samus charge shot
-        it_803F7288,
-        NULL,
-        it_2725_Logic108_Destroyed,
-        it_2725_Logic108_PickedUp,
-        NULL,
-        NULL,
-        itSamusChargeshot_Logic108_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic108_Reflected,
-        itSamusChargeshot_Logic108_Clanked,
-        itSamusChargeshot_Logic108_Absorbed,
-        it_2725_Logic108_ShieldBounced,
-        itSamusChargeshot_Logic108_HitShield,
-        itSamusChargeshot_Logic108_EvtUnk,
-    },
-    {
-        // Samus missile
-        it_803F7340,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic52_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic52_Reflected,
-        it_2725_Logic52_Clanked,
-        NULL,
-        it_2725_Logic52_ShieldBounced,
-        it_2725_Logic52_HitShield,
-        it_2725_Logic52_EvtUnk,
-    },
-    {
-        // Samus grapple beam
-        it_803F73A8,
-        itSamusGrapple_Logic53_Spawned,
-        NULL,
-        itSamusGrapple_Logic53_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itSamusGrapple_Logic53_EvtUnk,
-    },
-    {
-        // Sheik chain
-        it_803F7438,
-        NULL,
-        NULL,
-        it_2725_Logic54_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itSeakChain_Logic54_EvtUnk,
-    },
-    {
-        // Peach explode
-        it_803F7488,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itPeachExplode_Logic55_DmgDealt,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itPeachExplode_Logic55_EvtUnk,
-    },
-    {
-        // Peach turnip
-        it_803F74A8,
-        NULL,
-        itPeachTurnip_Logic56_Destroyed,
-        itPeachTurnip_Logic56_PickedUp,
-        itPeachTurnip_Logic56_Dropped,
-        itPeachTurnip_Logic56_Thrown,
-        itPeachTurnip_Logic56_DmgDealt,
-        NULL,
-        NULL,
-        itPeachTurnip_Logic56_Reflected,
-        itPeachTurnip_Logic56_Clanked,
-        NULL,
-        itPeachTurnip_Logic56_ShieldBounced,
-        itPeachTurnip_Logic56_HitShield,
-        itPeachTurnip_Logic56_EvtUnk,
-    },
-    {
-        // Bowser flame
-        ItemStateTable_KoopaFlame,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itKoopaFlame_Logic111_DmgDealt,
-        NULL,
-        NULL,
-        itKoopaFlame_Logic111_Reflected,
-        itKoopaFlame_Logic111_Clanked,
-        itKoopaFlame_Logic111_Absorbed,
-        itKoopaFlame_Logic111_ShieldBounced,
-        itKoopaFlame_Logic111_HitShield,
-        itKoopaFlame_Logic111_EvtUnk,
-    },
-    {
-        // Ness bat
-        it_803F6C68,
-        NULL,
-        NULL,
-        it_2725_Logic58_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itNessBat_Logic58_EvtUnk,
-    },
-    {
-        // Ness yo-yo
-        it_803F7558,
-        NULL,
-        NULL,
-        itNessYoyo_Logic59_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic59_EvtUnk,
-    },
-    {
-        // Peach parasol
-        it_803F74F8,
-        NULL,
-        itPeachParasol_Logic60_Destroyed,
-        itPeachParasol_Logic60_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itPeachParasol_Logic60_EvtUnk,
-    },
-    {
-        // Peach toad
-        it_803F7528,
-        NULL,
-        itPeachToad_Logic91_Destroyed,
-        itPeachToad_Logic91_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itPeachToad_Logic91_EvtUnk,
-    },
-    {
-        // Luigi fireball
-        it_803F75C0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itLuigiFireball_Logic89_DmgDealt,
-        NULL,
-        NULL,
-        itLuigiFireball_Logic89_Reflected,
-        itLuigiFireball_Logic89_Clanked,
-        itLuigiFireball_Logic89_Absorbed,
-        itLuigiFireball_Logic89_ShieldBounced,
-        itLuigiFireball_Logic89_HitShield,
-        itLuigiFireball_Logic89_EvtUnk,
-    },
-    {
-        // Ice Climbers ice
-        it_803F7668,
-        NULL,
-        it_2725_Logic90_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        itClimbersIce_Logic90_DmgDealt,
-        NULL,
-        NULL,
-        itClimbersIce_Logic90_Reflected,
-        itClimbersIce_Logic90_Clanked,
-        itClimbersIce_Logic90_Absorbed,
-        itClimbersIce_Logic90_ShieldBounced,
-        it_2725_Logic90_HitShield,
-        itClimbersIce_Logic90_EvtUnk,
-    },
-    {
-        // Ice Climbers blizzard
-        it_803F76A8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itClimbersBlizzard_DmgDealt,
-        NULL,
-        NULL,
-        itClimbersBlizzard_Reflected,
-        itClimbersBlizzard_Clanked,
-        itClimbersBlizzard_Absorbed,
-        itClimbersBlizzard_ShieldBounced,
-        itClimbersBlizzard_HitShield,
-        itClimbersBlizzard_EvtUnk,
-    },
-    {
-        // Zelda din fire
-        ItemStateTable_ZeldaDinFire,
-        NULL,
-        itZeldaDinFire_Logic65_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itZeldaDinFire_Logic65_Reflected,
-        itZeldaDinFire_Logic65_Clanked,
-        itZeldaDinFire_Logic65_Absorbed,
-        NULL,
-        NULL,
-        itZeldaDinFire_Logic65_EvtUnk,
-    },
-    {
-        // Zelda din fire explode
-        it_803F7740,
-        NULL,
-        itZeldaDinFireExplode_Logic66_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itZeldaDinFireExplode_Logic66_Clanked,
-        itZeldaDinFireExplode_Logic66_Absorbed,
-        itZeldaDinFireExplode_Logic66_ShieldBounced,
-        itZeldaDinFireExplode_Logic66_HitShield,
-        itZeldaDinFireExplode_Logic66_EvtUnk,
-    },
-    {
-        // Mewtwo disable
-        it_803F7750,
-        NULL,
-        itMewtwoDisable_Logic67_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        itMewtwoDisable_Logic67_DmgDealt,
-        NULL,
-        NULL,
-        itMewtwoDisable_Logic67_Reflected,
-        itMewtwoDisable_Logic67_Clanked,
-        itMewtwoDisable_Logic67_Absorbed,
-        itMewtwoDisable_Logic67_ShieldBounced,
-        itMewtwoDisable_Logic67_HitShield,
-        itMewtwoDisable_Logic67_EvtUnk,
-    },
-    {
-        // Peach toad spore
-        it_803F7548,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itPeachToadSpore_Logic92_DmgDealt,
-        NULL,
-        NULL,
-        itPeachToadSpore_Logic68_Reflected,
-        itPeachToadSpore_Logic68_Clanked,
-        itPeachToadSpore_Logic68_Absorbed,
-        itPeachToadSpore_Logic68_ShieldBounced,
-        itPeachToadSpore_Logic68_HitShield,
-        itPeachToadSpore_Logic92_EvtUnk,
-    },
-    {
-        // Mewtwo shadow ball
-        it_803F7760,
-        NULL,
-        it_2725_Logic101_Destroyed,
-        it_2725_Logic101_PickedUp,
-        NULL,
-        NULL,
-        itMewtwoShadowball_Logic101_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic101_Reflected,
-        itMewtwoShadowball_Logic101_Clanked,
-        itMewtwoShadowball_Logic101_Absorbed,
-        it_2725_Logic101_ShieldBounced,
-        itMewtwoShadowball_Logic101_HitShield,
-        itMewtwoShadowball_Logic101_EvtUnk,
-    },
-    {
-        // Ice Climbers string (belay)
-        it_803F76B8,
-        NULL,
-        NULL,
-        it_2725_Logic70_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic70_EvtUnk,
-    },
-    {
-        // Game & Watch greenhouse (insecticide spray)
-        it_803F7898,
-        NULL,
-        itGamewatchGreenhouse_Destroyed,
-        itGamewatchGreenhouse_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGamewatchGreenhouse_EvtUnk,
-    },
-    {
-        // Game & Watch manhole
-        it_803F78D8,
-        NULL,
-        itGameWatchManhole_Logic72_Destroyed,
-        itGameWatchManhole_Logic72_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchManhole_Logic72_EvtUnk,
-    },
-    {
-        // Game & Watch fire
-        it_803F78E8,
-        NULL,
-        itGamewatchFire_Destroyed,
-        itGamewatchFire_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGamewatchFire_EvtUnk,
-    },
-    {
-        // Game & Watch parachute
-        it_803F78F8,
-        NULL,
-        itGameWatchParachute_Logic74_Destroyed,
-        itGameWatchParachute_Logic74_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchParachute_Logic74_EvtUnk,
-    },
-    {
-        // Game & Watch turtle
-        it_803F7918,
-        NULL,
-        itGameWatchTurtle_Logic75_Destroyed,
-        itGameWatchTurtle_Logic75_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchTurtle_Logic75_EvtUnk,
-    },
-    {
-        // Game & Watch breath
-        it_803F7938,
-        NULL,
-        itGameWatchBreath_Logic76_Destroyed,
-        itGameWatchBreath_Logic76_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchBreath_Logic76_EvtUnk,
-    },
-    {
-        // Game & Watch judge
-        it_803F7968,
-        NULL,
-        itGameWatchJudge_Logic77_Destroyed,
-        it_2725_Logic77_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchJudge_Logic77_EvtUnk,
-    },
-    {
-        // Game & Watch panic
-        it_803F79A0,
-        NULL,
-        itGameWatchPanic_Logic78_Destroyed,
-        itGameWatchPanic_Logic78_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchPanic_Logic78_EvtUnk,
-    },
-    {
-        // Game & Watch chef
-        it_803F79E0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchChef_Logic112_DmgDealt,
-        NULL,
-        NULL,
-        itGameWatchChef_Logic112_Reflected,
-        it_2725_Logic112_Clanked,
-        it_2725_Logic112_Absorbed,
-        itGameWatchChef_Logic112_ShieldBounced,
-        it_2725_Logic112_HitShield,
-        itGameWatchChef_Logic112_EvtUnk,
-    },
-    {
-        // Young Link milk
-        it_803F7A28,
-        NULL,
-        NULL,
-        it_2725_Logic80_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itCLinkMilk_Logic80_EvtUnk,
-    },
-    {
-        // Game & Watch rescue
-        it_803F79C0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchRescue_Logic81_EvtUnk,
-    },
-    {
-        // Master Hand laser
-        it_803F9378,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itMasterHandLaser_Logic84_EvtUnk,
-    },
-    {
-        // Master Hand bullet
-        it_803F9388,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itMasterHandBullet_Logic83_DmgDealt,
-        NULL,
-        NULL,
-        itMasterHandBullet_Logic83_Reflected,
-        itMasterHandBullet_Logic83_Clanked,
-        itMasterHandBullet_Logic83_Absorbed,
-        itMasterHandBullet_Logic83_ShieldBounced,
-        itMasterHandBullet_Logic83_HitShield,
-        itMasterHandBullet_Logic85_EvtUnk,
-    },
-    {
-        // Crazy Hand laser
-        it_803F9378,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itMasterHandLaser_Logic84_EvtUnk,
-    },
-    {
-        // Crazy Hand bullet
-        it_803F9388,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itMasterHandBullet_Logic85_EvtUnk,
-    },
-    {
-        // Crazy Hand bomb
-        it_803F93A8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itCrazyHandBomb_Logic86_DmgDealt,
-        NULL,
-        NULL,
-        itCrazyHandBomb_Logic86_Reflected,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itCrazyHandBomb_Logic86_EvtUnk,
-    },
-    {
-        it_803F6788,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itMarioFireball_Logic87_DmgDealt,
-        NULL,
-        NULL,
-        itMarioFireball_Logic87_Reflected,
-        itMarioFireball_Logic87_Clanked,
-        itMarioFireball_Logic87_Absorbed,
-        itMarioFireball_Logic87_ShieldBounced,
-        itMarioFireball_Logic87_HitShield,
-        itMarioFireball_Logic87_EvtUnk,
-    },
-    {
-        it_803F75D0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itDrMarioPill_DmgDealt,
-        NULL,
-        NULL,
-        itDrMarioPill_Reflected,
-        itDrMarioPill_Clanked,
-        itDrMarioPill_Absorbed,
-        itDrMarioPill_ShieldBounced,
-        itDrMarioPill_HitShield,
-        itDrMarioPill_EvtUnk,
-    },
-    {
-        it_803F75C0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itLuigiFireball_Logic89_DmgDealt,
-        NULL,
-        NULL,
-        itLuigiFireball_Logic89_Reflected,
-        itLuigiFireball_Logic89_Clanked,
-        itLuigiFireball_Logic89_Absorbed,
-        itLuigiFireball_Logic89_ShieldBounced,
-        itLuigiFireball_Logic89_HitShield,
-        itLuigiFireball_Logic89_EvtUnk,
-    },
-    {
-        it_803F7668,
-        NULL,
-        it_2725_Logic90_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        itClimbersIce_Logic90_DmgDealt,
-        NULL,
-        NULL,
-        itClimbersIce_Logic90_Reflected,
-        itClimbersIce_Logic90_Clanked,
-        itClimbersIce_Logic90_Absorbed,
-        itClimbersIce_Logic90_ShieldBounced,
-        it_2725_Logic90_HitShield,
-        itClimbersIce_Logic90_EvtUnk,
-    },
-    {
-        it_803F7528,
-        NULL,
-        itPeachToad_Logic91_Destroyed,
-        itPeachToad_Logic91_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itPeachToad_Logic91_EvtUnk,
-    },
-    {
-        it_803F7548,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itPeachToadSpore_Logic92_DmgDealt,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itPeachToadSpore_Logic92_EvtUnk,
-    },
-    {
-        it_803F67D0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itFoxLaser_Logic94_Clanked,
-        NULL,
-        NULL,
-        itFoxLaser_Logic94_Reflected,
-        itFoxLaser_Logic94_Clanked,
-        itFoxLaser_Logic94_Absorbed,
-        itFoxLaser_Logic94_ShieldBounced,
-        itFoxLaser_Logic94_HitShield,
-        itFoxLaser_Logic94_EvtUnk,
-    },
-    {
-        it_803F67D0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itFoxLaser_Logic94_Clanked,
-        NULL,
-        NULL,
-        itFoxLaser_Logic94_Reflected,
-        itFoxLaser_Logic94_Clanked,
-        itFoxLaser_Logic94_Absorbed,
-        itFoxLaser_Logic94_ShieldBounced,
-        itFoxLaser_Logic94_HitShield,
-        itFoxLaser_Logic94_EvtUnk,
-    },
-    {
-        it_803F6CA8,
-        NULL,
-        NULL,
-        itFoxBlaster_Logic96_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itFoxBlaster_Logic96_EvtUnk,
-    },
-    {
-        it_803F6CA8,
-        NULL,
-        NULL,
-        itFoxBlaster_Logic96_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itFoxBlaster_Logic96_EvtUnk,
-    },
-    {
-        it_803F6A28,
-        NULL,
-        itLinkArrow_Logic98_Destroyed,
-        itLinkArrow_Logic98_PickedUp,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_DmgDealt,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_Reflected,
-        itLinkArrow_Logic98_Clanked,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_HitShield,
-        itLinkArrow_Logic98_EvtUnk,
-    },
-    {
-        it_803F6A28,
-        NULL,
-        itLinkArrow_Logic98_Destroyed,
-        itLinkArrow_Logic98_PickedUp,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_DmgDealt,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_Reflected,
-        itLinkArrow_Logic98_Clanked,
-        NULL,
-        NULL,
-        itLinkArrow_Logic98_HitShield,
-        itLinkArrow_Logic98_EvtUnk,
-    },
-    {
-        it_803F6E98,
-        NULL,
-        itLinkBow_Logic100_Destroyed,
-        itLinkBow_Logic100_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itLinkBow_Logic100_EvtUnk,
-    },
-    {
-        it_803F6E98,
-        NULL,
-        itLinkBow_Logic100_Destroyed,
-        itLinkBow_Logic100_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itLinkBow_Logic100_EvtUnk,
-    },
-    {
-        it_803F7760,
-        NULL,
-        it_2725_Logic101_Destroyed,
-        it_2725_Logic101_PickedUp,
-        NULL,
-        NULL,
-        itMewtwoShadowball_Logic101_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic101_Reflected,
-        itMewtwoShadowball_Logic101_Clanked,
-        itMewtwoShadowball_Logic101_Absorbed,
-        it_2725_Logic101_ShieldBounced,
-        itMewtwoShadowball_Logic101_HitShield,
-        itMewtwoShadowball_Logic101_EvtUnk,
-    },
-    {
-        it_803F6B70,
-        NULL,
-        it_2725_Logic102_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itNesspkflash_Logic102_Reflected,
-        itNesspkflash_Logic102_Clanked,
-        itNesspkflash_Logic102_Absorbed,
-        NULL,
-        NULL,
-        itNesspkflash_Logic102_EvtUnk,
-    },
-    {
-        it_803F6F40,
-        NULL,
-        it_2725_Logic103_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itNessPKFlashExplode_Logic103_Clanked,
-        itNessPKFlashExplode_Logic103_Absorbed,
-        itNessPKFlashExplode_Logic103_ShieldBounced,
-        itNessPKFlashExplode_Logic103_HitShield,
-        itNessPKFlashExplode_Logic103_EvtUnk,
-    },
-    {
-        it_803F7190,
-        NULL,
-        it_2725_Logic106_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic106_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic106_Reflected,
-        it_2725_Logic106_Clanked,
-        it_2725_Logic106_Absorbed,
-        it_2725_Logic106_ShieldBounced,
-        it_2725_Logic106_HitShield,
-        itPikachuTJoltGround_Logic106_EvtUnk,
-    },
-    {
-        it_803F71D8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic107_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic107_Reflected,
-        it_2725_Logic107_Clanked,
-        it_2725_Logic107_Absorbed,
-        it_2725_Logic107_ShieldBounced,
-        it_2725_Logic107_HitShield,
-        itPikachuTJoltAir_Logic107_EvtUnk,
-    },
-    {
-        it_803F7190,
-        NULL,
-        it_2725_Logic106_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic106_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic106_Reflected,
-        it_2725_Logic106_Clanked,
-        it_2725_Logic106_Absorbed,
-        it_2725_Logic106_ShieldBounced,
-        it_2725_Logic106_HitShield,
-        itPikachuTJoltGround_Logic106_EvtUnk,
-    },
-    {
-        it_803F71D8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic107_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic107_Reflected,
-        it_2725_Logic107_Clanked,
-        it_2725_Logic107_Absorbed,
-        it_2725_Logic107_ShieldBounced,
-        it_2725_Logic107_HitShield,
-        itPikachuTJoltAir_Logic107_EvtUnk,
-    },
-    {
-        it_803F7288,
-        NULL,
-        it_2725_Logic108_Destroyed,
-        it_2725_Logic108_PickedUp,
-        NULL,
-        NULL,
-        itSamusChargeshot_Logic108_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic108_Reflected,
-        itSamusChargeshot_Logic108_Clanked,
-        itSamusChargeshot_Logic108_Absorbed,
-        it_2725_Logic108_ShieldBounced,
-        itSamusChargeshot_Logic108_HitShield,
-        itSamusChargeshot_Logic108_EvtUnk,
-    },
-    {
-        it_803F6F50,
-        NULL,
-        itSeakNeedleThrown_Logic109_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic109_DmgDealt,
-        it_2725_Logic109_DmgReceived,
-        NULL,
-        it_2725_Logic109_Reflected,
-        it_2725_Logic109_Clanked,
-        NULL,
-        it_2725_Logic109_ShieldBounced,
-        it_2725_Logic109_HitShield,
-        itSeakNeedleThrown_Logic109_EvtUnk,
-    },
-    {
-        it_803F70A8,
-        NULL,
-        NULL,
-        itSeakNeedleHeld_Logic110_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itSeakNeedleHeld_Logic110_EvtUnk,
-    },
-    {
-        ItemStateTable_KoopaFlame,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itKoopaFlame_Logic111_DmgDealt,
-        NULL,
-        NULL,
-        itKoopaFlame_Logic111_Reflected,
-        itKoopaFlame_Logic111_Clanked,
-        itKoopaFlame_Logic111_Absorbed,
-        itKoopaFlame_Logic111_ShieldBounced,
-        itKoopaFlame_Logic111_HitShield,
-        itKoopaFlame_Logic111_EvtUnk,
-    },
-    {
-        it_803F79E0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGameWatchChef_Logic112_DmgDealt,
-        NULL,
-        NULL,
-        itGameWatchChef_Logic112_Reflected,
-        it_2725_Logic112_Clanked,
-        it_2725_Logic112_Absorbed,
-        itGameWatchChef_Logic112_ShieldBounced,
-        it_2725_Logic112_HitShield,
-        itGameWatchChef_Logic112_EvtUnk,
-    },
-    {
-        it_803F7958,
-        NULL,
-        itKirbyGameWatchChefPan_Logic113_Destroyed,
-        itKirbyGameWatchChefPan_Logic113_PickedUp,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itKirbyGameWatchChefPan_Logic113_EvtUnk,
-    },
-    {
-        it_803F94A0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_27CF_Logic114_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_27CF_Logic114_EvtUnk,
-    },
-    {
-        // Unk4
-        it_803F8BD8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2E5A_Logic115_DmgDealt,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2E5A_Logic115_EvtUnk,
-    },
-    {
-        // Coin
-        it_803F93C8,
-        itCoin_Logic116_Spawned,
-        itCoin_Logic116_Destroyed,
-        itCoin_Logic116_PickedUp,
-        itCoin_Logic116_Thrown,
-        itCoin_Logic116_Thrown,
-        NULL,
-        itCoin_Logic116_DmgReceived,
-        itCoin_Logic116_EnteredAir,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itCoin_Logic116_EvtUnk,
-    },
-    {
-        // Pokemon random
-        it_803F8C8C,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2E6A_Logic117_DmgDealt,
-        it_2E6A_Logic117_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2E6A_Logic117_EvtUnk,
-    },
-};
-
-/// Monster and Stage-specific Items
-struct sdata_ItemGXLink it_803F4CA8[30] = {
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC,
-    it_8026EECC, it_8026EECC, it_8026EECC, it_8026EECC, NULL,
-};
-
-struct ItemLogicTable it_803F4D20[ARRAY_SIZE(it_803F4CA8)] = {
-    {
-        // Old Goomba (old-Kuri)
-        it_803F8320,
-        it_802D73F0,
-        it_2725_Logic9_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic0_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802D84D8,
-    },
-    {
-        // Target (Mato)
-        it_803F83E0,
-        it_802D84F8,
-        it_2725_Logic9_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802D85F4,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Yoshi's Story Shy Guy (Heiho)
-        it_803F83F0,
-        it_802D8688,
-        it_2725_Logic9_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802D8EC8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802D9A0C,
-    },
-    {
-        // Koopa Troopa (Green) (Nokonoko)
-        it_803F85F0,
-        it_802DC69C,
-        it_2725_Logic9_Destroyed,
-        it_802DD4F4,
-        it_802DD59C,
-        it_802DD67C,
-        NULL,
-        itNokonoko_Logic3_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802DD7D0,
-    },
-    {
-        // Koopa Troopa (Red) (Patapata)
-        it_803F88C0,
-        it_802E05A0,
-        it_2725_Logic9_Destroyed,
-        itPatapata_Logic4_PickedUp,
-        it_2725_Logic4_Dropped,
-        it_2725_Logic4_Thrown,
-        NULL,
-        it_2725_Logic4_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802E16D8,
-    },
-    {
-        // Like-Like (likelike)
-        it_803F8468,
-        it_802D9BA8,
-        itLikeLike_Logic5_Destroyed,
-        itLikeLike_Logic5_PickedUp,
-        it_2725_Logic5_Dropped,
-        itLikeLike_Logic5_Thrown,
-        NULL,
-        it_2725_Logic5_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802E16D8,
-    },
-    {
-        // Old Redead (old-lead) [invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Old Octorok (old-octa) [invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Old Ottosea (old-otto)
-        it_803F89C8,
-        it_802E2470,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic8_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Polar Bear (whitebea)
-        it_803F8A88,
-        it_802E37BC,
-        it_2725_Logic9_Destroyed,
-        itWhiteBea_Logic9_PickedUp,
-        it_2725_Logic9_Dropped,
-        it_2725_Logic9_Thrown,
-        NULL,
-        it_802E3884,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802E4A24,
-    },
-    {
-        // Klaptrap (klap)
-        it_803F8940,
-        it_802E1820,
-        itKlap_Logic10_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_2725_Logic10_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802E2450,
-    },
-    {
-        // Green Shell (zgshell)
-        it_803F86C8,
-        it_802DE040,
-        itZGShell_Logic11_Destroyed,
-        it_2725_Logic11_PickedUp,
-        itZGShell_Logic11_Dropped,
-        itZGShell_Logic11_Thrown,
-        itZGShell_Logic11_DmgDealt,
-        itZGShell_Logic11_DmgReceived,
-        it_2725_Logic11_EnteredAir,
-        itZGShell_Logic11_Reflected,
-        it_2725_Logic11_Clanked,
-        NULL,
-        itZGShell_Logic11_ShieldBounced,
-        it_2725_Logic11_HitShield,
-        it_802DFED4,
-    },
-    {
-        // Red Shell (green act) (zrshell)
-        it_803F87F0,
-        it_802E02E8,
-        itZRShell_Logic12_Destroyed,
-        itZRShell_Logic12_PickedUp,
-        itZRShell_Logic12_Dropped,
-        itZRShell_Logic12_Thrown,
-        itZRShell_Logic12_DmgDealt,
-        itZRShell_Logic12_DmgReceived,
-        itZRShell_Logic12_EnteredAir,
-        itZRShell_Logic12_Reflected,
-        itZRShell_Logic12_Clanked,
-        NULL,
-        itZRShell_Logic12_ShieldBounced,
-        itZRShell_Logic12_HitShield,
-        it_802E0468,
-    },
-    {
-        // Tingle (Tincle) (on balloon)
-        it_803F9000,
-        it_802EB5C8,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itTincle_Logic13_DmgDealt,
-        itTincle_Logic13_DmgReceived,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802EC850,
-    },
-    {
-        // [Invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // [Invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // [Invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Whispy Apple
-        it_803F91D0,
-        NULL,
-        NULL,
-        it_802EEA08,
-        it_802EEA70,
-        it_802EEB28,
-        it_802EEED0,
-        NULL,
-        itWhispyApple_Logic18_EnteredAir,
-        it_802EEF10,
-        it_802EEED8,
-        NULL,
-        it_802EEF68,
-        it_802EEF30,
-        it_802EEF88,
-    },
-    {
-        // Whispy's Healing Apple
-        it_803F91D0,
-        NULL,
-        NULL,
-        it_802EEA08,
-        it_802EEA70,
-        it_802EEB28,
-        it_802EEED0,
-        NULL,
-        itWhispyApple_Logic18_EnteredAir,
-        it_802EEF10,
-        it_802EEED8,
-        NULL,
-        it_802EEF68,
-        it_802EEF30,
-        it_802EEF88,
-    },
-    {
-        // [Invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // [Invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // [Invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Tool (Flatzone)
-        it_803F9260,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itTools_Logic22_DmgDealt,
-        itTools_Logic22_DmgReceived,
-        NULL,
-        itTools_Logic22_Reflected,
-        it_2725_Logic22_Clanked,
-        it_2725_Logic22_Absorbed,
-        itTools_Logic22_ShieldBounced,
-        it_2725_Logic22_HitShield,
-        it_802EFA24,
-    },
-    {
-        // [Invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // [Invalid]
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
-    {
-        // Birdo (Kyasarin)
-        it_803F90D0,
-        it_802ECE90,
-        itKyasarin_Logic25_Destroyed,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802EDDC0,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        it_802EE1E0,
-    },
-    {
-        // Arwing Laser
-        it_803F8DE8,
-        NULL,
-        it_802E79C8,
-        NULL,
-        NULL,
-        NULL,
-        it_802E838C,
-        NULL,
-        NULL,
-        it_802E8420,
-        it_802E838C,
-        it_802E8418,
-        it_802E85F4,
-        it_802E8784,
-        NULL,
-    },
-    {
-        // Great Fox's Laser
-        it_803F8FD0,
-        it_802EB1EC,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        itGreatFoxLaser_Logic27_DmgDealt,
-        NULL,
-        NULL,
-        it_2725_Logic27_Reflected,
-        itGreatFoxLaser_Logic27_Clanked,
-        itGreatFoxLaser_Logic27_Absorbed,
-        NULL,
-        NULL,
-        it_802EB5A8,
-    },
-    {
-        // Birdo's Egg
-        it_803F9328,
-        NULL,
-        NULL,
-        itKyasarinEgg_Logic28_PickedUp,
-        itKyasarinEgg_Logic28_Dropped,
-        itKyasarinEgg_Logic28_Thrown,
-        it_2725_Logic28_DmgDealt,
-        it_802F022C,
-        NULL,
-        it_2725_Logic28_Reflected,
-        it_2725_Logic28_Clanked,
-        NULL,
-        itKyasarinEgg_Logic28_ShieldBounced,
-        it_2725_Logic28_HitShield,
-        it_802F0320,
-    },
-    {
-        // NULL
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-        NULL,
-    },
 };
 
 typedef enum PokemonKind {
