+++ src/melee/it/it_3F2F.h
@@ -1,0 +1,12 @@
+#ifndef GALE01_IT_3F2F
+#define GALE01_IT_3F2F
+
+#include "it/items/types.h"
+#include "it/types.h"
+
+/* 3F2F28 */ extern struct sdata_ItemGXLink it_803F2F28[118];
+/* 3F3100 */ extern struct ItemLogicTable it_803F3100[118];
+/* 3F4CA8 */ extern struct sdata_ItemGXLink it_803F4CA8[30];
+/* 3F4D20 */ extern struct ItemLogicTable it_803F4D20[30];
+
+#endif
